Changes in taylor

+++ Chapter1/taylor_exp_eval.cpp
@@ -2,7 +2,7 @@
 #include<cmath>
 #include<cstdlib>
 
-int factorial( int n )
+long int factorial( int n )
 {
     return (n == 1 || n == 0) ? 1 : factorial(n-1)*n;
 }
@@ -10,7 +10,7 @@ int factorial( int n )
 int main( int argc, char* argv[] )
 {
     int n = std::atoi(argv[2]);
-    double exp_x, x = std::atof(argv[1]);
+    long double exp_x, x = std::atof(argv[1]);
     for ( int i = 0; i < n; i++ )
 	exp_x += pow( x, i )/factorial(i);
